moved toJson functionality out of Entity
